res insert transition ani

+++ errSrv/apiRes.go
@@ -2,8 +2,11 @@ package errSrv
 
 import (
 	"bytes"
+	"github.com/disintegration/imaging"
+	"github.com/h2non/filetype"
 	"github.com/kataras/iris/v12"
 	"gorm.io/gorm/clause"
+	"log"
 	"math/rand"
 	"mime/multipart"
 	"os"
@@ -15,7 +18,8 @@ import (
 const maxSize = 8 * iris.MB
 
 var fileCache map[string][]multipart.File
-var fileInfoCache map[string][2]string
+var fileInfoCache map[string][3]string
+var fileFirstCache map[string][]byte
 var chs map[int64]chan string
 
 func initResApi(app *iris.Application) {
@@ -42,10 +46,23 @@ func resCh(ctx iris.Context) {
 
 func resDel(ctx iris.Context) {
 	id := ctx.URLParam("id")
-	ids := strings.Split(id, ",")
-	err := db.Delete(&Res{}, "ID in ?", ids).Error
-	countRes()
-	handleErr(ctx, err)
+	if id == "" {
+		ctx.StatusCode(200)
+	} else {
+		ids := strings.Split(id, ",")
+		err := db.Delete(&Res{}, "ID in ?", ids).Error
+		countRes()
+		handleErr(ctx, err)
+		addJob(func() {
+			for _, i := range ids {
+				er := os.Remove("./dist/" + i)
+				_ = os.Remove("./dist/" + i + ".png")
+				if er != nil {
+					log.Printf("del %s file fail: %v \n", i, er)
+				}
+			}
+		})
+	}
 }
 func resLs(ctx iris.Context) {
 	pg := ctx.Params().GetIntDefault("page", 1)
@@ -79,22 +96,30 @@ func upload(ctx iris.Context) {
 	ctx.SetMaxRequestBodySize(maxSize)
 	key := ctx.FormValue("title")
 	nm := ctx.FormValue("name")
-	tp := ctx.FormValue("type")
 	if fileCache == nil {
 		fileCache = make(map[string][]multipart.File)
-		fileInfoCache = make(map[string][2]string)
-	}
-	if nm != "" {
-		fileInfoCache[key] = [2]string{nm, tp}
+		fileInfoCache = make(map[string][3]string)
+		fileFirstCache = make(map[string][]byte)
 	}
 	pt := ctx.FormValue("part")
 	part, _ := strconv.Atoi(pt)
 	total, _ := strconv.Atoi(ctx.FormValue("total"))
+	total = total - 1
 	if _, ok := fileCache[key]; !ok {
 		fileCache[key] = make([]multipart.File, total)
 	}
 	file, _, _ := ctx.FormFile("data")
-	fileCache[key][part] = file
+
+	if pt == "0" {
+		buf := new(bytes.Buffer)
+		_, _ = buf.ReadFrom(file)
+		bt := buf.Bytes()
+		kind, _ := filetype.Match(bt)
+		fileInfoCache[key] = [3]string{nm, kind.MIME.Value, kind.Extension}
+		fileFirstCache[key] = bt
+	} else {
+		fileCache[key][part-1] = file
+	}
 	done := true
 	for i := 0; i < total; i++ {
 		if fileCache[key][i] == nil {
@@ -103,43 +128,63 @@ func upload(ctx iris.Context) {
 		}
 	}
 	if done {
-		_ = os.Mkdir("dist", os.ModePerm)
-		fn := "dist/" + key
-		f, _ := os.OpenFile(fn, os.O_WRONLY|os.O_CREATE, 0666)
-		defer func(f *os.File) {
-			err := f.Close()
-			if err != nil {
+		addJob(func() {
+			_ = os.Mkdir("dist", os.ModePerm)
+			fn := "dist/" + key
+			f, _ := os.OpenFile(fn, os.O_WRONLY|os.O_CREATE, 0666)
+			defer func(f *os.File) {
+				err := f.Close()
+				if err != nil {
+				}
+			}(f)
+			fk := fileFirstCache[key]
+			che := fileCache[key]
+			ii, er := f.Write(fk)
+			i := int64(ii)
+			for _, ff := range che {
+				buf := new(bytes.Buffer)
+				_, _ = buf.ReadFrom(ff)
+				s, _ := f.Seek(i, 0)
+				ii, er = f.WriteAt(buf.Bytes(), s)
+				if er != nil {
+					log.Printf("write file err %v", er)
+				}
+				i = int64(ii) + i
 			}
-		}(f)
-		i := int64(0)
-		for _, ff := range fileCache[key] {
-			buf := new(bytes.Buffer)
-			_, _ = buf.ReadFrom(ff)
-			s, _ := f.Seek(i, 0)
-			ii, _ := f.WriteAt(buf.Bytes(), s)
-			i = int64(ii) + i
+			inf, _ := fileInfoCache[key]
+			re := Res{
+				Name: inf[0],
+				Type: inf[1],
+				Ext:  inf[2],
+				Size: i,
+				ID:   key,
+				Date: time.Now().Unix(),
+			}
+			db.Clauses(clause.OnConflict{
+				UpdateAll: true,
+			}).Create(re)
+			delete(fileCache, key)
+			delete(fileInfoCache, key)
+			delete(fileFirstCache, key)
+			if strings.HasPrefix(re.Type, "image") {
+				thumbnail(f)
+			}
+			for _, cha := range chs {
+				if cha != nil {
+					cha <- strings.Join([]string{key, pt}, ",")
+				}
+			}
+		})
+	} else {
+		for _, cha := range chs {
+			if cha != nil {
+				nf := fileInfoCache[key]
+				cha <- strings.Join([]string{key, pt, nf[1], nf[2]}, ",")
+			}
 		}
-		inf, _ := fileInfoCache[key]
-		re := Res{
-			Name: inf[0],
-			Type: inf[1],
-			Size: i,
-			ID:   key,
-			Date: time.Now().Unix(),
-		}
-		db.Clauses(clause.OnConflict{
-			UpdateAll: true,
-		}).Create(re)
-		delete(fileCache, key)
-		delete(fileInfoCache, key)
 	}
-	countRes()
 	ctx.StatusCode(200)
-	for _, cha := range chs {
-		if cha != nil {
-			cha <- strings.Join([]string{key, pt}, ",")
-		}
-	}
+	countRes()
 }
 
 func msg(ctx iris.Context) {
@@ -162,6 +207,8 @@ func msg(ctx iris.Context) {
 			go func() {
 				_ = <-cha
 				close(cha)
+				cha = nil
+				delete(chs, key)
 			}()
 			time.Sleep(time.Millisecond * 10)
 			cha = nil
@@ -171,5 +218,18 @@ func msg(ctx iris.Context) {
 		c := <-cha
 		_, _ = ctx.Writef("data: %s\n\n", c)
 		flusher.Flush()
+	}
+}
+
+func thumbnail(f *os.File) {
+	n := f.Name()
+	_ = f.Close()
+	img, err := imaging.Open(n)
+	if err == nil {
+		img := imaging.Resize(img, 0, 200, imaging.Lanczos)
+		err = imaging.Save(img, n+".png")
+		if err != nil {
+			log.Printf("thumbnail err %v", err)
+		}
 	}
 }
